Restore gd32 as dominant target
Catchup on old trees. Prepare to land disk. Some work for multiple boards (e.g. gd32-dev) in build.

+++ usr/sys/h/map.h
@@ -3,7 +3,8 @@
 struct map
 {
 	short	m_size;
-	unsigned short m_addr;
+	// unsigned short m_addr;
+	unsigned long m_addr;
 };
 
 struct map coremap[CMAPSIZ];	/* space for core allocation */
